requirements.txt
asgiref==3.5.2
crispy-bootstrap5==0.6
Django==3.2.15
django-crispy-forms==1.14.0
django-environ==0.9.0
Pillow==9.2.0
pytz==2022.2.1
sqlparse==0.4.2
